// apps/api/src/modules/devoluciones/domain/entities/devolucion-enums.ts
export enum EstadoDevolucion {
  SOLICITADA = 'SOLICITADA',
  EN_REVISION = 'EN_REVISION',
  APROBADA = 'APROBADA',
  RECHAZADA = 'RECHAZADA',
  COMPLETADA = 'COMPLETADA',
}

export enum ResolucionDevolucion {
  REEMBOLSO = 'REEMBOLSO',
  CAMBIO = 'CAMBIO',
  SIN_COMPENSACION = 'SIN_COMPENSACION',
}

export enum DestinoFisicoItem {
  INVENTARIO_RESTOCK = 'INVENTARIO_RESTOCK',
  BAJA_DANADO = 'BAJA_DANADO',
  INSPECCION = 'INSPECCION',
}